fixes #31 #32
