+++ app/src/main/AndroidManifest.xml
@@ -39,29 +39,6 @@
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="com.android.news24x7.activities.HomeActivity" />
         </activity>
-        <!-- SyncAdapter's dummy authentication service
-        <service android:name=".sync.NewsAuthenticatorService">
-            <intent-filter>
-                <action android:name="android.accounts.AccountAuthenticator" />
-            </intent-filter>
-            <meta-data
-                android:name="android.accounts.AccountAuthenticator"
-                android:resource="@xml/authenticator" />
-        </service>
-
-         The SyncAdapter service
-    <service
-            android:name=".sync.NewsSyncService"
-            android:exported="true"
-            >
-            <intent-filter>
-                <action android:name="android.content.SyncAdapter" />
-            </intent-filter>
-            <meta-data
-                android:name="android.content.SyncAdapter"
-                android:resource="@xml/syncadapter" />
-        </service>-
-    This is required for immediate syncs -->
     <service
         android:name=".sync.NewsSyncIntentService"
         android:exported="false" />
