--- app/src/main/java/com/android/news24x7/FABAwareScrollingViewBehavior.java
@@ -1,59 +1,0 @@
-package com.android.news24x7;
-
-import android.content.Context;
-import android.support.design.widget.AppBarLayout;
-import android.support.design.widget.CoordinatorLayout;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v4.view.ViewCompat;
-import android.util.AttributeSet;
-import android.view.View;
-
-import java.util.List;
-
-/**
- * Created by Dell on 5/23/2017.
- */
-
-public class FABAwareScrollingViewBehavior extends AppBarLayout.ScrollingViewBehavior {
-    public FABAwareScrollingViewBehavior(Context context, AttributeSet attrs) {
-        super(context, attrs);
-    }
-
-    @Override
-    public boolean layoutDependsOn(CoordinatorLayout parent, View child, View dependency) {
-        return super.layoutDependsOn(parent, child, dependency) ||
-                dependency instanceof FloatingActionButton;
-    }
-
-    @Override
-    public boolean onStartNestedScroll(final CoordinatorLayout coordinatorLayout, final View child,
-                                       final View directTargetChild, final View target, final int nestedScrollAxes) {
-        // Ensure we react to vertical scrolling
-        return nestedScrollAxes == ViewCompat.SCROLL_AXIS_VERTICAL
-                || super.onStartNestedScroll(coordinatorLayout, child, directTargetChild, target, nestedScrollAxes);
-    }
-
-    @Override
-    public void onNestedScroll(final CoordinatorLayout coordinatorLayout, final View child,
-                               final View target, final int dxConsumed, final int dyConsumed,
-                               final int dxUnconsumed, final int dyUnconsumed) {
-        super.onNestedScroll(coordinatorLayout, child, target, dxConsumed, dyConsumed, dxUnconsumed, dyUnconsumed);
-        if (dyConsumed > 0) {
-            // User scrolled down -> hide the FAB
-            List<View> dependencies = coordinatorLayout.getDependencies(child);
-            for (View view : dependencies) {
-                if (view instanceof FloatingActionButton) {
-                    ((FloatingActionButton) view).hide();
-                }
-            }
-        } else if (dyConsumed < 0) {
-            // User scrolled up -> show the FAB
-            List<View> dependencies = coordinatorLayout.getDependencies(child);
-            for (View view : dependencies) {
-                if (view instanceof FloatingActionButton) {
-                    ((FloatingActionButton) view).show();
-                }
-            }
-        }
-    }
-}
