+++ app/src/main/java/com/android/news24x7/sync/NewsSyncTask.java
@@ -10,9 +10,11 @@ import android.util.Log;
 import com.android.news24x7.BuildConfig;
 import com.android.news24x7.database.NewsUtil;
 import com.android.news24x7.parcelable.Article;
+import com.android.news24x7.prefs.NewsPreferences;
 import com.android.news24x7.retrofit.ApiClient;
 import com.android.news24x7.retrofit.ApiInterface;
 import com.android.news24x7.retrofit.NewsResponse;
+import com.android.news24x7.util.NotificationUtils;
 
 import java.util.ArrayList;
 import java.util.HashMap;
