+++ app/src/main/java/com/android/news24x7/sync/NewsSyncUtils.java
@@ -23,7 +23,7 @@ import java.util.concurrent.TimeUnit;
 public class NewsSyncUtils {
 
 
-    private static final int REMINDER_INTERVAL_MINUTES = 1;
+    private static final int REMINDER_INTERVAL_MINUTES = 30;
     private static final int REMINDER_INTERVAL_SECONDS = (int) (TimeUnit.MINUTES.toSeconds(REMINDER_INTERVAL_MINUTES));
     private static final int SYNC_FLEXTIME_SECONDS = REMINDER_INTERVAL_SECONDS;
 
@@ -68,7 +68,6 @@ public class NewsSyncUtils {
                     startImmediateSync(context);
                 }
 
-               cursor.close();
             }
         });
