+++ app/src/main/res/layout/fragment_article_detail.xml
@@ -45,7 +45,7 @@
         android:id="@+id/scrollview"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        app:layout_behavior="com.android.news24x7.FABAwareScrollingViewBehavior">
+        app:layout_behavior="com.android.news24x7.interfaces.FABAwareScrollingViewBehavior">
 
         <LinearLayout
             android:layout_width="match_parent"
